Fix referencing the fields from the Matrix4f class

+++ src/org/joml/Vector3d.java
@@ -1085,9 +1085,9 @@ public class Vector3d implements Externalizable {
      * @return dest
      */
     public Vector3d mulPosition(Matrix4f mat, Vector3d dest) {
-        dest.set(mat.ms[Matrix4d.M00] * x + mat.ms[Matrix4d.M10] * y + mat.ms[Matrix4d.M20] * z + mat.ms[Matrix4d.M30],
-                 mat.ms[Matrix4d.M01] * x + mat.ms[Matrix4d.M11] * y + mat.ms[Matrix4d.M21] * z + mat.ms[Matrix4d.M31],
-                 mat.ms[Matrix4d.M02] * x + mat.ms[Matrix4d.M12] * y + mat.ms[Matrix4d.M22] * z + mat.ms[Matrix4d.M32]);
+        dest.set(mat.ms[Matrix4f.M00] * x + mat.ms[Matrix4f.M10] * y + mat.ms[Matrix4f.M20] * z + mat.ms[Matrix4f.M30],
+                 mat.ms[Matrix4f.M01] * x + mat.ms[Matrix4f.M11] * y + mat.ms[Matrix4f.M21] * z + mat.ms[Matrix4f.M31],
+                 mat.ms[Matrix4f.M02] * x + mat.ms[Matrix4f.M12] * y + mat.ms[Matrix4f.M22] * z + mat.ms[Matrix4f.M32]);
         return dest;
     }
 
@@ -1118,10 +1118,10 @@ public class Vector3d implements Externalizable {
      * @return the <i>w</i> component of the resulting 4D vector after multiplication
      */
     public double mulPositionW(Matrix4f mat, Vector3d dest) {
-        double w = mat.ms[Matrix4d.M03] * x + mat.ms[Matrix4d.M13] * y + mat.ms[Matrix4d.M23] * z + mat.ms[Matrix4d.M33];
-        dest.set(mat.ms[Matrix4d.M00] * x + mat.ms[Matrix4d.M10] * y + mat.ms[Matrix4d.M20] * z + mat.ms[Matrix4d.M30],
-                 mat.ms[Matrix4d.M01] * x + mat.ms[Matrix4d.M11] * y + mat.ms[Matrix4d.M21] * z + mat.ms[Matrix4d.M31],
-                 mat.ms[Matrix4d.M02] * x + mat.ms[Matrix4d.M12] * y + mat.ms[Matrix4d.M22] * z + mat.ms[Matrix4d.M32]);
+        double w = mat.ms[Matrix4f.M03] * x + mat.ms[Matrix4f.M13] * y + mat.ms[Matrix4f.M23] * z + mat.ms[Matrix4f.M33];
+        dest.set(mat.ms[Matrix4f.M00] * x + mat.ms[Matrix4f.M10] * y + mat.ms[Matrix4f.M20] * z + mat.ms[Matrix4f.M30],
+                 mat.ms[Matrix4f.M01] * x + mat.ms[Matrix4f.M11] * y + mat.ms[Matrix4f.M21] * z + mat.ms[Matrix4f.M31],
+                 mat.ms[Matrix4f.M02] * x + mat.ms[Matrix4f.M12] * y + mat.ms[Matrix4f.M22] * z + mat.ms[Matrix4f.M32]);
         return w;
     }
